requirements.txt
streamlit
googletrans==4.0.0rc1
gTTS
openai-whisper
torch  # Required for whisper
